Entrega

+++ apps/biblioteca/views.py
@@ -46,24 +46,26 @@ def nova_imagem(request):
 
     return render(request, 'galeria/nova_imagem.html', {'form': form})
 
-def editar_imagem(request, foto_id):   
-    fotografia = Livro.objects.get(id=foto_id)
-    form = FotografiaForms(instance=fotografia)
+def editar_imagem(request, foto_id):
+    if request.user.is_staff:
+        fotografia = Livro.objects.get(id=foto_id)
+        form = FotografiaForms(instance=fotografia)
 
-    if request.method == 'POST':
-        form = FotografiaForms(request.POST, request.FILES, instance=fotografia)
-        if form.is_valid():
-            form.save()
-            messages.success(request, "Livro alterado!")
-            return redirect('imagem', foto_id)
+        if request.method == 'POST':
+            form = FotografiaForms(request.POST, request.FILES, instance=fotografia)
+            if form.is_valid():
+                form.save()
+                messages.success(request, "Livro alterado!")
+                return redirect('imagem', foto_id)
 
     return render(request, 'galeria/editar_imagem.html', {'form': form, 'foto_id': foto_id})
 
 def deletar_imagem(request, foto_id):
-    fotografia = Livro.objects.get(id=foto_id)
-    fotografia.delete()
-    messages.success(request, "Deleção feita com sucesso!")
-    return redirect('index')
+    if request.user.is_staff:
+        fotografia = Livro.objects.get(id=foto_id)
+        fotografia.delete()
+        messages.success(request, "Deleção feita com sucesso!")
+        return redirect('index')
 
 def filtro(request, categoria):
     fotografias = Livro.objects.order_by('id').filter(publicada=True, categoria=categoria)
